Moved all context-cache modules from lib/context to a library of its own - to prepare for persisting the context cache in mongo
