+++ src/lib/orionld/context/orionldContextPresent.cpp
@@ -32,7 +32,7 @@ extern "C"
 
 #include "orionld/context/OrionldContext.h"                      // OrionldContext
 #include "orionld/context/OrionldContextItem.h"                  // OrionldContextItem
-#include "orionld/context/orionldContextCache.h"                 // ORIONLD_CONTEXT_CACHE_HASH_ARRAY_SIZE
+#include "orionld/contextCache/orionldContextCache.h"            // ORIONLD_CONTEXT_CACHE_HASH_ARRAY_SIZE
 #include "orionld/context/orionldContextPresent.h"               // Own interface
