+++ src/lib/orionld/contextCache/orionldContextCacheInit.cpp
@@ -22,20 +22,24 @@
 *
 * Author: Ken Zangelin
 */
-#include <unistd.h>                                              // NULL
-#include <semaphore.h>                                           // sem_t
+#include <strings.h>                                             // bzero
 
-#include "orionld/context/OrionldContext.h"                      // OrionldContext
-#include "orionld/context/orionldContextCache.h"                 // Own interface
+#include "logMsg/logMsg.h"                                       // LM_*
+#include "logMsg/traceLevels.h"                                  // Lmt*
+
+#include "orionld/contextCache/orionldContextCache.h"            // Context Cache Internals
+#include "orionld/contextCache/orionldContextCacheInit.h"        // Own interface
 
 
 
 // -----------------------------------------------------------------------------
 //
-// Context Cache Internals
+// orionldContextCacheInit -
 //
-sem_t             orionldContextCacheSem;
-OrionldContext*   orionldContextCacheArray[100];  // When 100 is not enough, a realloc is done (automatically)
-OrionldContext**  orionldContextCache         = orionldContextCacheArray;
-int               orionldContextCacheSlots    = 100;
-int               orionldContextCacheSlotIx   = 0;
+void orionldContextCacheInit(void)
+{
+  bzero(&orionldContextCacheArray, sizeof(orionldContextCacheArray));
+
+  if (sem_init(&orionldContextCacheSem, 0, 1) == -1)
+    LM_X(1, ("Runtime Error (error initializing semaphore for orionld context list; %s)", strerror(errno)));
+}
